Introduce `thiserror`
Previously, the error type was `String`. For libraries, this is not ideal,
because it is hard to match on the error as a downstream user.
Furthermore, it requires to use conversions like `.map_err(|err| err.to_string())`.
This change introduces [`thiserror`](https://github.com/dtolnay/thiserror),
which should make error handling ergonomic moving forward.